feat: add refresh request parameters management to OAuth2 component and remove deprecated files
